C Programming in Linux Tutorial

+++ 0x15-file_io/0-read_textfile.c
@@ -40,7 +40,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	buffer[bytes_read] = '\0';
 
 	bytes_written = write(STDOUT_FILENO, buffer, bytes_read);
-	if (bytes_written == -1 || (size_t)bytes_written != bytes_read)
+	if (bytes_written == -1 || bytes_written != bytes_read)
 	{
 		close(fd);
 		free(buffer);
